fixed termination

+++ Instance.py
@@ -1,5 +1,5 @@
 '''
-Created on Sep 3, 2013
+Created on Sep 9, 2013
 
 @author: eviek.
 '''
@@ -28,17 +28,21 @@ class InstanceHandler:
         self.instances = []
     
     def launch(self):
+        print "Perform configurations in AWS ec2 and launch instances.."
         conn = boto.ec2.connect_to_region(self.region, aws_access_key_id=self.aws_access_key_id,
                                           aws_secret_access_key=self.aws_secret_access_key)
         # Create private key and save to disk
-        keypair = conn.create_key_pair(self.key)
-        keypair.save("/tmp")
+#        keypair = conn.create_key_pair(self.key)
+#        keypair.save("/tmp")
         # Create simple security group to allow ssh connections
-        self.group = conn.create_security_group(name=self.security_group, description=self.security_group)
-        self.group.authorize('tcp', 22, 22, '0.0.0.0/0')
+#        self.group = conn.create_security_group(name=self.security_group, description=self.security_group)
+#        self.group.authorize('tcp', 22, 22, '0.0.0.0/0')
         # Launch instances
+#        reservation = conn.run_instances(self.image, min_count=self.count, max_count=self.count,
+#                                         key_name=self.key, security_groups=[self.group], 
+#                                         instance_type=self.instance_type, placement=self.availabity_zone)
         reservation = conn.run_instances(self.image, min_count=self.count, max_count=self.count,
-                                         key_name=self.key, security_groups=[self.group], 
+                                         key_name=self.key, security_groups=[self.security_group], 
                                          instance_type=self.instance_type, placement=self.availabity_zone)
         # Wait for all instances to be up and running
         self.instances = reservation.instances
@@ -60,7 +64,7 @@ class InstanceHandler:
             print name + "\t" + instance.id + "\t" + instance.public_dns_name
             hosts.append(instance.public_dns_name)
             ips.append(instance.private_ip_address)
-        print 'Private key path: /tmp/' + self.key + '.pem'
+#        print 'Private key path: /tmp/' + self.key + '.pem'
         conn.close()
         # Return list of hostnames
         return hosts, ips
@@ -68,32 +72,20 @@ class InstanceHandler:
     def terminate(self):
         conn = boto.ec2.connect_to_region(self.region, aws_access_key_id=self.aws_access_key_id,
                                           aws_secret_access_key=self.aws_secret_access_key)
-        if not self.instances:
-            # Get reservations with running instances
-            reservations = conn.get_all_instances(filters={'instance-state-name': 'running'})
-            if reservations:
-                for r in reservations:
-                    self.instances = self.instances + r.instances
-                sys.stdout.write("Running instances: ")
-                sys.stdout.write(self.instances[0].id)
-                for instance in self.instances[1:]:
-                    sys.stdout.write(", " + instance.id)
-                sys.stdout.write("\nTerminate all running instances? [Y/n] ")
-                answer = raw_input().lower()
-                while answer not in ["", "y", "ye", "yes", "n", "no"]:
-                    print "Please respond with 'yes'/'no' (or 'y'/'n')."
-                    sys.stdout.write("Terminate all running instances? [Y/n] ")
-                    answer = raw_input().lower()
-                if answer in ["n", "no"]:
-                    self.instances = []
-                    print 'Aborted.'
         if self.instances:
+            print "Terminating instances.."
             # Create a list of instances' ids
             term = []
             for instance in self.instances:
                 term.append(instance.id)
             # Terminate instances
             terminated = conn.terminate_instances(term)
+            # Wait for all instances to completely shut-down
+            for instance in terminated:
+                status = instance.update()
+                while status == 'shutting-down':
+                    time.sleep(2)
+                    status = instance.update()
             print "Successfully terminated instances: ",
             sys.stdout.write(terminated[0].id)
             for instance in terminated[1:]:
@@ -107,36 +99,58 @@ class InstanceHandler:
                 failed = [i for i in term if i not in s]
                 for instance in failed:
                     print instance
-            # Wait for all instances to completely shut-down
-            for instance in self.instances:
-                status = instance.update()
-                while status == 'shutting-down':
-                    time.sleep(2)
-                    status = instance.update()
-            try:
+#            try:
                 # Delete created key and security group
-                conn.delete_key_pair(self.key)
-                conn.delete_security_group(self.security_group)
-                os.remove(str('/tmp/' + self.key + '.pem'))
-                print 'Removed keypair from disk'
-            except:
-                print "There are remaining keypairs and/or security groups. Make sure to delete"
-                print "them if not needed."
+#                conn.delete_key_pair(self.key)
+#                conn.delete_security_group(self.security_group)
+#                os.remove(str('/tmp/' + self.key + '.pem'))
+#                print 'Removed keypair from disk'
+#            except:
+#                print "There are remaining keypairs and/or security groups. Make sure to delete"
+#                print "them if not needed."
         else:
             print "No running instances to terminate"
         conn.close()
+        
+    def get_instances(self):
+        hosts = []
+        ips = []
+        conn = boto.ec2.connect_to_region(self.region, aws_access_key_id=self.aws_access_key_id,
+                                          aws_secret_access_key=self.aws_secret_access_key)
+        for i in range(2, len(sys.argv)):
+            res = conn.get_all_instances(filters={'instance-id': str(sys.argv[i])})
+            if res:
+                self.instances = self.instances + res[0].instances
+                hosts.append(res[0].instances[0].public_dns_name)
+                ips.append(res[0].instances[0].private_ip_address)
+            else:
+                print "Instance " + str(sys.argv[i]) + " does not exist"
+                sys.stdout.write("Are you sure you want to continue? [Y/n] ")
+                answer = raw_input().lower()
+                while answer not in ["", "y", "ye", "yes", "n", "no"]:
+                    print "Please respond with 'yes'/'no' (or 'y'/'n')."
+                    sys.stdout.write("Are you sure you want to continue? [Y/n] ")
+                    answer = raw_input().lower()
+                if answer in ["n", "no"]:
+                    self.instances = []
+                    hosts = []
+                    ips = []
+                    print 'Aborted.'
+                    break
+        conn.close()
+        return hosts, ips
 
 if __name__ == "__main__":    
     instances = InstanceHandler()
-    if len(sys.argv) == 2:
-        if 'launch' == sys.argv[1]:
-            instances.launch()
-        elif 'terminate' == sys.argv[1]:
-            instances.terminate()
-        else:
-            print "Unknown command"
-            sys.exit(2)
-        sys.exit(0)
-    else:
-        print "usage: %s launch|terminate" % sys.argv[0]
+    if len(sys.argv) == 2 and 'launch' == sys.argv[1]:
+        instances.launch()
+    elif len(sys.argv) > 2 and 'terminate' == sys.argv[1]:
+        instances.get_instances()
+        instances.terminate()
+    elif 'terminate' != sys.argv[1] and 'launch' != sys.argv[1]:
+        print "Unknown command"
         sys.exit(2)
+    else:
+        print "usage: %s launch|terminate instance1 instance2 .." % sys.argv[0]
+        sys.exit(2)
+    sys.exit(0)
